cmd_area_mex.lua widget fixes and enhancements
- Use "ctrl" key for split build
- Enhanced mex placement (tries closest first)
Known issues:
==> Mex placement can be resource hungry for large areas with multiple
constructors (needs further optimizations)
==> Metal detection isn't absolute and might fail on some maps (mostly
maps where Mex Snap also fails)
==> The limit to "how far from the center of a spot will it allow to
place a mex" is set to be equal to extractorRadius when it should be
(extractorRadius - spotsize). In some situations mexes might work at non
optimal efficiency (because something was in the way and it was put
further rather then discarded).