Scavenger | Final fixes/tweaks/balance
We're almost ready for like a first official 'release'.
We've done loads of bug fixing, tweaking and balancing.

There still are many ideas in the works, but for now it's (almost) a full AI.
Things we've planned are more balanced spawning and/or some little favor in attacking winning teams, etc.