High trajectory fix and balance adjustments to plasma defences
The armguard, corpun, armamb, and cortoast will now properly transition from low to high trajectory and back.  Guardian and Punisher have their special damages removed and their aoe reduced, but have +20% default damage.  The special damages for the ambusher and toaster are removed.  The toaster's damage was buffed so that the unit is no longer strictly worse than the ambusher.