Sea balance adjustments
Frigates given a 10% health buff so that they are better at surviving skirmishes.  Hovertanks given a speed buff to help make them better at raiding and more difficult to kite.  Snapper buffed to be on par with the anaconda.  Supporter cost reduced slightly to be more evenly balanced with decade.  Fast attack hovers given more reasonable buildtimes.  Halberd given a los that is more in line with the rest of the hovercraft.