Add ancillary LuaUI files
These files are included by gameside logic (see main.lua:20 and
barwidgets.lua:22 plus the following lines). However they are not
present in the repo, so are always user local (this usually works
because the engine also distributes them, but delete your local
LuaUI folder to see what happens otherwise).

Adding them to the game itself means two things:

* idiot-proofing. These files are crucial for LuaUI to load
  so it no longer becomes possible to screw up your install
  by deleting them locally.

* enables hax-proofing. In the future, local user widgets will
  have limited access to some interfaces, but this could be
  worked around via modified versions of these files (which are
  loaded quite early). Having them in the game means we can
  enforce VFS to load them. Doing that is not in the scope of
  this patch however (currently local versions are preferred,
  despite the game-side versions now existing).

These files are from 2007 and the quality of their design and
implementation leaves something to be desired. Improving it
is also not in the scope of this patch.