Update cmd_bar_hotkeys.lua
N was missing from the keybinding table for layout support. Attempts to bind to N would have resulted in crashing the widget. Discovered while modifying the widget with a different keybind layout so I thought I should put a PR here to fix it in case someone changes the keybinds in this one and adds a bind for N.