fix(smartselect): Deselects when reaching minimap
Fix an issue where I assumed incorrectly that x1, y1 was origin vertex
(the one that is fixed) on selectbox, causing deselections to happen
when reaching minimap.

Consolidate two loops to 2n -> n and also check for equality of
modifiers and selections to return early and avoid unnecessary
computation.

It is important to note that we are avoiding returning early on
identical coordinates and same modifiers (as before) since we want to
select units that move into the selection box, e.g. patrolling etc.

We also ignore the engine selection and reselect our own, without
passing the SelectionChanged to widgets on this run. This incurs in one
additional selection but we both avoid the previous expensive cache
check and ensure correctness.