Add explosionSpeed to commanderexplosion
comupdate explosion changes made the commander explosion prorogate at a finite rate. At the edge, the damage from the explosion could be 18 frames delayed. 

This broke the mo_preventcombomb gadget because UnitPreDamaged could not get the attackerID of the commander that died many frame ago, and could not check the last commander condition.