Balance changes to T1 missile trucks and T1 veh artillery
Whistler range decreased 600 ->575 as it was too easy to kite enemies, especially bots.  Lasher was not as strong at kiting due to its slower speed, but an early group can still apply a lot of pressure so its damage is being reduced by 6%.  This change also helps to differentiate each missile truck to match its faction's playstyle.  The T1 vehicle artillery (Shellshocker and Wolverine) are also having their effective dps reduced by 7% in order to give the defender more time to respond to an artillery push.