luarules: hotfix `game_ffa_start_setup` due to missing engine release
In 6bb4e64fd80cb357e3c721b05b6b63e44290aea4, we reworked FFA start
points and additionally added a shuffling system for start boxes.

However, this shuffling requires `Spring.SetAllyTeamStartBox` to be
available in engine, and no release has been made available with it yet.
This results in a crash at startup for FFA games:

```
[string "LuaRules/Gadgets/game_ffa_start_setup.lua"]:154: attempt to call field 'SetAllyTeamStartBox' (a nil value)) ptop=3 cenv=false vfsmode=Mmb
```

As a hotfix, we temporarily comment out the start box shuffling code. It
will be re-enabled later when a new engine release is out.