fix: gui_reclaim_field_highlight: fix broken refactor in UpdateFeatures (#2345)
Commit 3b27aaf6cfdb67fa01e4a8f5989c564e3b755cb3 changed UpdateFeatures
to iterate over knownFeatures instead of Spring.GetAllFeatures(). But
these have different structures (Spring.GetAllFeatures() is a list of
featureIDs, knownFeatures is a dictionary of featureID -> featureInfo).

The result was that this part of UpdateFeatures didn't work at all, and
feature locations and metal content were only updated when a feature was
created/destroyed.

This fix changes the iteration to match the different data structure.