fix(gui_options): use correct function in attackrange_gl4 options spec (#2400)
Previously, the value for attackrange_cursorunitrange was set from the
return value of setCursorUnitRange(), not getCursorUnitRange(). This led
to "C stack overflow" errors if setCursorUnitRange() was called from
here, presumably because setCursorUnitRange() calls widget:Initialize().
In addition, the attackrange_gl4 widget would then be disabled.

In most cases, this wasn't an issue, because options are usually
initialized before other widgets are loaded. This meant that the nil
check on setCursorUnitRange always failed, and short-circuited the call.

But if a user widget adds a setting, gui_options.Init() is called again
after widgets have been loaded, which means that setCursorUnitRange is
not nil, and is called, leading to the error.