fix(api_resource_spot_builder): don't issue commands if justDraw==true (#2396)
9e85deb6101a69f5460f03bf08aa4c2fd81c43ea refactored a bunch of this
code, and inadvertently moved the code to issue guard commands to
secondary builders outside of a justDraw check.

One result was that when t1 and t2 constructors were selected and a
metal spot was hovered, cmd_rclick_quick_build_resource_extractor would
call BuildMex with justDraw==true, and a guard command would be issued
every frame. This caused flickering and repeated sound effects in
addition to the unintended orders.