fix(gui_selfd_icons): correctly handle removing queued commands (#2509)
If there was a command queue ({MOVE,SELFD}, for example), then executing
a REMOVE on the MOVE command would leave the widget thinking that the
queue had been replaced and there was no longer a SELFD command in the
queue.

This change adds a check for commands that don't necessarily clear the
queue, even when executed without shift. The check is not necessarily
exhaustive - these are only the ones I'm aware of.