feat(gui_building_grid_gl4): show grid on water surface when needed (#2716)
Previously, the grid would always follow the heightmap. This works fine
for buildings that are built on the ground (above or below water). But
for floating buildings, the grid would be misaligned from the actual
building location.

This change adds a check to see if the building is underwater, in which
cases it uses the previous behavior. Otherwise, it causes the grid go no
lower than the water level.