fix(instancevbotable): use unique instanceIDs (#2826)
* style(instancevbotable): remove extra whitespace

* fix(instancevbotable): use unique instanceIDs

Previously, in pushElementInstance, new instanceIDs were allocated based
on the number of instances stored. This can lead to situations where,
after elements are removed and added, and updateExisting is used, an
existing instanceID is allocated to a new instance.

This shortcut was also present in popElementInstance, where it allowed
implicitly removing the last element by not providing an instanceID.
This can lead to similar issues as in pushElementInstance, in that it
uses the number-of-instances-stored as an instanceID.

This commit fixes this issue by always issuing unique (within a given
table) instanceIDs within pushElementInstance, and removing the
nil-instanceID feature from popElementInstance.