Gadget: Drone Carriers, Bot Launchers, and Unit Builders inherit/bequeath XP (#2932)
New customparams added that allows units to share xp to the unit that created it, and for a unit to give its xp to units it builds.  The relative power level math is handled within the gadget.  Used currently for evolving commanders, drone carriers, and a few extra units like the armbotrail.