EvoCom patch 8 (#2975)
* fix raptor queen cheesable via dgun

raptor queen couldn't kill Cortex commander and sometimes even Armada commander before being D-gunned to death, damage resistance wasn't enough to counteract this. This patch aims to fix that.

* armada commander EMP tumbleweeds  higher priority than commander

This is to help Armada commander retreat.

* fixed incorrect objectname when xmas happens

glad I beat christmas, because this model wouldn't load if it happened

* EMP tumbleweed has energycost now

this is to prevent errors from doing multiplication functions with a power value of 0 when xp is gained.

* reduce range/damage of Cortex Commander sea laser

now it's same as Armada range, and damage reduced to compensate for the fact the shield blocks torpedos.

* balance cortex shields

they now aren't the most goddang powerful thing in the whole stinkin' game. Feels more balnaced to me.

* rebalanced all sea weapons for commanders

the original DPS values were based on land confrontations. Sea units generally have less DPS and less health than land units for the metal, so the commanders were just mopping the floor with them. It's now far more reasonable.

* set combat distance evolution delay to 0

this caused problems where the commander evolutions would become desynced due to some being engaged in combat when combat evolution triggered.

The players voted for this delay to be removed.

* add nano turret build options to legion com

the players voted for this

* armada commander can shoot while cloaked

the players voted for this. The poll suggested a -50% nerf to HP to compensate, but i felt that was too punishing due to the ubiquity of radar coverage.  It's -33% hp instead relative to other commanders starting at level 3

* make commander eco scaling more linear

this follows player poll results

* fix evolution time inconsistency

it's overwritten anyways, but I like consistency in my unitdefs

* add arm cannon to cortex commander, nerf flamethrower to compensate

* fix indenting in alldefs