replace hardcoded armor type indices (#3013)
Removed the hardcoded damages[14] for vtol damage in GUI. Replaced with a table of indices, eg damages[armors.vtol].

This solves an issue since the evocom update which added a new armortype (probably PR#2975) named 'raptorqueen'. Array positions of armor types is alphabetical, so vtol was moved by 1. Units are displaying their damage to subs instead of to vtol since ~yesterday.