Update unit_carrier_spawner.lua
Added support for evolving carriers. When a carrier evolves it will now transfer drones to the evolved unit. If the evolved unit is no longer a carrier, the drones will still be transferred, but docking will be disabled.

Fixed a bug causing disabling of docking not to work as intended.