Legion Hovercraft Update
Legion now has their own hovercraft plant with 4 new hovercraft, models and scripts made by EnderRobo.  The Glaucus fills the role of the hoverscout with a reduced cost and speed compared to other hoverscouts.  Nereus is the hovertank, with a lower cost and higher speed than the other hovertanks and a double cannon, but a reduced health.  Salacia is the rocket hovercraft.  Alpheus is an anti-air hovercraft with an aa minigun.