fix(gridmenu,info): A few minor cases (#3227)
* feat(gridmenu): Select first item available

Instead of first in the list, if it's not available pick the next
available when autoselecting on category switch

* fix(info): Fix special activecmd coloring

It was never being successful since exposed selectedID does not exist by
buildmenu. Instead we use the activecmd, which should have been used in
the first place.

* fix(gridmenu): Stale gui_info

Info panel would get stale when hovering a cell and switching selection
without moving the mouse to a non constructor unit (e.g. control group)

* fix(gridmenu): Defensively check areamex widget

If areamex broke, gridmenu should not break

* Just a small added safety check

* Forgot to check this guy

* fix(info): Not showing pregame build selection