add modoption no_comtrans (#3126)
At the time of making this commit, a tweakunit setting disallowing t1
transports from carrying Armada and Cortex commanders has been decently
popular among map rotation lobby players. The downside of using the
tweakunit is that players need to remember the exact base64 string to
enable it and to be able to distinguish that it is said tweakunit that
is being used. Moreover, the tweakunit setting does not tweak Legion
commanders.

To improve quality of life for players, add a new modoption no_comtrans
that sets mass of all commanders to 5001 (rather than 4900 or 4999).