Armada evocom rework (#3480)
Half rework of the armada evocom. Now it's more oriented towards porc turtling than hit and run stuff. Included some bugfixes in here too for effigies and a little balancing stuff for legion and cortex.

Also a complete rewrite of the workertime_boost gadget I made a while ago. It was way too big of a performance impact for the tiny job it was doing. Now it also includes several other categories of things that can be used to trigger workertime boosts as well. Performance now clocks under 0.001% down from 1.0%ish

Ran this through a full round of scavengers to make sure there's no nil errors.