Add Epic Commando, Rework Paratrooper into a customParam (#3494)
Tested as I went. The epic commando is intended to be a scavenger-only unit that sneaks intot he back line and blows up your base if you don't find it in time.

Reworked paratrooper gadget to be a customparam while I'm at it. Dunno what the hat business was about so I left that piece of code as is.