Add fall damage, fix effigy transports (#3597)
enables the fall damage gadget and creates some exceptions for a rulesparam added for respawn effigies. Now when a transport is destroyed while carrying a unit that can respawn, it won't be destroyed but instead will fall. If it dies from the fall damage, it will respawn.

I tried triggering respawn upon transport destruction and had weird results. It seems we can't change the position of a unit while they're falling using Spring.SetUnitPosition. As such, this is the best solution for now.

While I was at it I made fall damage a customparam and gave the commandos 0 fall damage. There is no changes made to vanilla gameplay in effect.