Feature: Factory Quotas (#3688)
Adds functionality that lets you set certain quotas for each unit type in the factory build menu, which are unit minimums. This means that the factory will queue new units of this type until the quota is fulfilled. These quotas are per-factory. The way you can set the quotas is by toggling a toggle, that tells the build menu to set quotas instead of setting the queue. The toggle will reset to setting the queue every time the selection is changed. The widget will use alt to queue the units.

Priority is handled like this (low numbers = high priority):

    alt-queued units (including ones from the quota widget)
    units from factory quotas
    a. quotas that have the most units missing from the quota (in percentage)
    b. quotas that have the least units missing from the quota
    normally queued units

The toggle is bound to factoryqueuemode, and stop production will also cancel quotas.