Add T3 Shields "Asylum" and "Sanctuary" (#4104)
![image](https://github.com/user-attachments/assets/25a94ff3-18a3-4bff-bf85-79308957fa9a)

Thanks to Protar for his modeling genius

Adds t3 shields Armada "Conservator" and the Cortex "Overlord" to the extra units pack. **Hides modoption to override default shield behavior.**

Sanctuary:
roughly 5 times as powerful and 5 times as costly as a normal cortex shield with +275 range.
blocks/deflects (depending on rework) all weapons except:
- large starburst missiles
- aircraft bombs
- lightening
- (until merge of my other [shield re-rework pr](https://github.com/beyond-all-reason/Beyond-All-Reason/pull/4077) ) flames

Asylum:
roughly 4 times as powerful and 4 times as costly as normal armada shield. +160 range.
blocks/deflects (depending on rework) all weapons except:
- large starburst missiles
- aircraft bombs
- flames

Legion is given access to the Conservator until their t2 and t3 shields are complete.

the opacity of the shields is a little stronger than their t2 counterparts. I'd like to change the tint but that isn't supported in the unitdefs and I wasn't able to figure out how to change it in the fx

**testing steps**

- enable extra units pack,
- you should now see build options for the t3 shields
- build them
- test all projectiles, note the above exclusions
- you can spawn them with `corgatet3` and `armgatet3`

*inversely*

- ensure nothing changes in vanilla game when extra units isn't enabled

Q/A
**But WHY?**
The general sentiment is we need fewer modoptions. Removing shield behavior overrides will cause riots without sweetening the deal a bit. This is that sugar, baby.

**Porc bad!**
Extra units is almost exclusively used for PVE games. In PVE, the AI doesn't really care if you're Porc'ing, and weak shields make porc hard in some of the more crazy game modes.

**Why the exceptions to specific projectiles?!**
So the shields aren't without counterplay. Starburst-style missiles and bombing runs are a deliberate choice. Flames/lightening is for faction flavor. Small missiles are blocked because of the catapult and because there exist many general assault rocketbots that would otherwise effortlessly demolish defenses in scavengers gamemode.