Smart High/low weapon selection proof of concept (#3720)
This implement's the GDT decision to remove toggles for high/low trajectory from the game. It accomplishes this by automating that process with a combination of gadget and .bos header that has been integrated into the animation scripts of the high/low units. This header doesn't just work for high/low, but for automated switching between a preferred weapon mode and a fallback weapon mode when the preferred weapon fails to aim or fire upon a target.