legion unit update (#4339)
This is the culmination of over 10 months of learning to work with
Blender and learning to work on BAR as a whole. The Legion Sol Invictus
is the halo tier 3 unit for the faction, serving as a door-busting
game-ending tool akin to the Juggernaut. I have wanted to make this unit
a reality for about as long as I have known about Legion's existence
within BAR, but never had the skills or knowledge to actualise upon it
until now. Everything I have modelled, scripted, and otherwise created
for BAR has led to this.

Specifically for this unit, there was extensive design and concept work
done by @Protar1458 to help get a much clearer and practical idea of
what is and isn't possible, and some of the modelling work is done by
him directly such as the head and the arms. This unit wouldn't have
existed in the awesome form as it does now without Protar's guiding hand
and skillful digital pen. As always in the spirit of iterative
improvements that BAR embodies, there is an upcoming update for this
unit that improves its legs and subsequent walking animations, also
attributed to Protar.

On top of that, @Zecrus2 has been an invaluable and irreplaceable source
of assistance and straight up esoteric knowledge bank to help get this
gargantuan unit behaving, aiming, and destroying units in the way it was
intended to. This unit would straight up not exist without his help.

Along the way, multiple other people have helped bestow tonnes of
knowlege on me to get me where I am, I can't possibly name them all but
chiefly Tharsis (without whose animation tutorial this probably would
have taken an extra month lol), EnderRobo, ChrisFloofyKitsune (without
whose awesome S3OKit addon for Blender the workflow would have taken yet
another month), Hornet, Hobo Joe, and Flaka.

This is my love letter to this game. Enjoy!

### Work Done
Adds the Sol Invictus, a late-t3 ultra-heavy mech.

### Work Needed
Future PR will improve legs and associated animations.

### Images
No peeking. Spin up a Legion lobby and have fun.

---------

Co-authored-by: Zecrus2 <72325649+Zecrus2@users.noreply.github.com>
Co-authored-by: Arbsquid <171824403+Arbsquid@users.noreply.github.com>
Co-authored-by: Protar <25440671+Protar1458@users.noreply.github.com>