feat(spotlight): add API for drawing spotlights on the map (#4350)
This exposes two functions:
`WG["ObjectSpotlight"].addSpotlight`
`WG["ObjectSpotlight"].removeSpotlight`

Each function takes a type ("unit", "feature", "ground"), an owner (usually the name of the widget), and an ID or XYZ position.

In addition, addSpotlight takes a color, as well as an options object, which can be used to change the height and radius, and also set an expiration time for the spotlight.