fix(pve): adding logs, boss spawning can return nil (#4505)
There is a a bug where the game does not end. ~~I have only seen it in
one out of three games so maybe rare.
[Replay](https://bar-rts.com/replays/ba4fd46755318df16ec88fca5b0bc2d9)~~
It seems to be quite common for large lobbies. What happens is probably
that some queen fails one or more spawns due to so much stuff on the
map. but the spawn counter is increased anyway preventing game over.
I'm not sure if this change solves it.

### Work done
* Added logs for spawning and killing bosses as well as setting a
rulesparam.
* Spawn function can return nil so checking for that before updating the
counters
* the `>=` shouldn't be needed but I don't see how it can hurt

#### Test steps
- [x] Tested raptor/scav defense in skirmish from start to game over
with 7 bosses