Fix scav and raptor finding of initial spawn locations (#4619)
### Work done

- Adjust scav and raptor startbox when smaller than twice the burrow
spread size so the area will always be searchable.
- Incrementally enlarge the search area when failing to create the
initial (first and second) burrows/beacons.
- Create a library for common scav and raptor code
(gadgets/include/SpawnerEnemyLib.lua)

#### Addresses Issue(s)
- Fixes
https://github.com/beyond-all-reason/Beyond-All-Reason/issues/4048
- Queen/boss spawning directly when spawn area too small and/or can't
find a place for the initial two burrows/beacons.

### Remarks

- The issues happen when the player chooses a too small startbox for the
scav/raptor player.
- Not finding a burrow location can also happen when the startbox
doesn't have enough flat areas.
- Talked to @Damgam about this, but maybe we can refine it further, or
change the solution a bit.
- Created the library because I see a lot of duplicated scav and raptor
code (and didn't want to add the new code duplicated), I think a lot of
the scav and raptor code can be moved here for better reuse and
maintainability, but let me know if you don't like that or think a
different place could be used. I see there's also damgam_lib, but I
think that one can be used for generic utilities, more than scav and
raptor code.
- didn't want to refactor a lot of existing code to not complicate the
PR more than needed, so for now I placed the purely new methods in the
library, and still had to perform some similar changes and additions for
scav and raptor files.

### Details

The algorithm works as follows

- First makes sure startbox width and height is bigger than twice
`spread = burrowsize*1.5`, since then we will be looking between
min+spread and max-spread, so anything smaller than that will fail
trying to find a random location inside (and generate lua errors when
trying to `random(min+spread, max-spread)`, since `min+spread >
max-spread`).
- After the first try, if it couldn't find a location for a spawn, it
will increase the margin to three times the spread, then four and so on.
This way the AI should be able to find a location, even in the face of
difficult terrain.
- Each spawn try is done every 30 frames, just like it was doing before,
the difference is it's enlarging the search area now (and also making
sure its initially big enough).

After the first burrow, we assume we already have a placeable area, but
still need to place a second burrow, so the code keeps trying with a
less aggressive startbox expansion algo:

- It will divide the remaining grace time by some number (currently 20),
and use that as max retries before incrementing startbox again.
- Doing it like this to let the random placer some tries to find
something, otherwise assume need to expand.
- If the grace period is arriving then that means we need to hurry up,
so it starts more aggressively expanding the area, until in the end
every second is going to be expanding.

### How to test

- Start a skirmish game against raptors or scav, and give the
raptor/scav player a very small startbox, or placed in an unspawnable
area.
- Start the game and see how the ai eventually finds a location.
  - also no more math.random errors.
- queen shouldn't spawn directly now since it should always manage a
first burrow spawn.

### Screenshots

As an example, when faced with a small box, that's also in an
unspawnable part of the map like this:


![raptorboxes](https://github.com/user-attachments/assets/3924db54-6f68-49c4-8553-d2bf8bdd44c5)

it will successfully spawn as seen in the following image:


![raptorsstartsm](https://github.com/user-attachments/assets/f09439ac-78b6-4445-b50d-7c430616e202)

example of where the second burrow is placed:


![raptors2sm](https://github.com/user-attachments/assets/aa281e60-7598-4f25-a1d4-7b787a39bb65)

similarly for scavs:


![scavs](https://github.com/user-attachments/assets/5e1ca8a2-a73c-43ef-bffd-7ced7dff579e)