Change extractor snap to respect "InvertQueueKey" (#4754)
"InvertQueueKey" allows players to avoid holding shift to queue items. Crucially, this is important for a keyboardless interface, where the `/invqueuekey` command allows players to toggle queueing without the use of a keyboard.

This is already supported in engine queueing, but when a user was trying to make use of this feature, they discovered that snapping buildings do not respect the "InvertQueueKey" option as other options in the build menu do.