Settarget with shift keybind file fix (#4930)
Currently, trying to do settarget via keybind with shift held isn't possible. This can be an issue if units already have stuff in their targeting queues and a player wants to add additional targets to the queues. The player has to let go of shift, press the key for setting targets, and then press shift again before adding additional targets.

This issue is only for settarget, and not any other of the similar commands (repair, attack etc.), which all have "double-binds" with and without shift.