Revert "Increasing performance of gui_unit_wait_icons.lua (way less m… (#5083)
…emory loads, if units are out of sight) (#5080)"

This reverts commit dd4dd2705f96bccff77169bb897aae86a3bd7f78.

<!--
PR Template! Please make sure to give your PR a relevant title so a
squash merge remains descriptive
If any commented sections are not relevant to this PR, remove them.
Please fill out the uncommented sections with any relevant information.
-->

### Work done
<!--
Describe the changes or additions made in this PR, and why they
are necessary or important. If there is unusual complexity in the
code or functionality, please explain it so reviewers can understand.
-->

<!-- If relevant
#### Addresses Issue(s)
- Issue URL
-->

<!-- If relevant
#### Setup
Describe any setup requirements to test this work (Specific settings,
widgets, etc))
-->

#### Test steps
- [ ] Write the steps needed to test out the changes. Include the
expected result.

<!-- If relevant
### Screenshots:
If you're making visible changes, add before/after screenshots or videos
of the major
changes so it's easier for reviewers to see what is different in this PR

#### BEFORE:
(screenshot from master)

#### AFTER:
(screenshot from branch)
-->