Legion T2 Air May 2025 Update (#5124)
### Work Done
Adds the remaining models for placeholder aircraft in the T2 Legion
Aircraft Plant: the advanced construction aircraft, the Whisper
Radar/Sonar Scout plane, and the Legionnaire defensive fighter (now
known as the Nemesis to avoid verbal confusion with "legion air").

The Wildfire will not have a model replacement, as it is subject to
removal once Faction Wars is over.

With this, Legion's T2 air roster is fully complete.

### Images

![{405BADD2-A2E4-4CA0-886D-80F24631B56F}](https://github.com/user-attachments/assets/3a475cc1-8e81-4d3a-918a-7e5c3f17b27d)

![{BD6EE970-BB05-4172-930D-A941D8496903}](https://github.com/user-attachments/assets/ea8e4100-5b94-4621-b60f-f386e99d23e4)