Acid and napalm anti-stacking (#4996)
This is a temporary fix to stacked damage areas. Tested it in game and feel like it strikes a good balance while patching this unwanted outcome.

The limit is a "soft cap" to prevent an enormous performance cliff right at the max stack count, something no player would be able to observe atm. This is not the same as limiting the number of "stacks" of various areas of effect that can damage a unit, but it is anti-stacking in a way.

Large damage values, above the per-second damage limit, partially bypass the limiter, to avoid penalizing the Perdition. Values otherwise above the limit are penalized by 80%, so heavily, though only in the portion that is above the threshold.