Faction limiter mk2 (#4922)
Adds the faction locker, implemented using a bitmask instead of text, as well as easier to use Chobby UI that doesn't require Spads regex.

Hidden in singleplayer unless the modoption is set to something else under _DEV tab or Show Hidden Modoptions is enabled. In Multiplayer it is shown when the player is the BOSS of the lobby.

When the modoption is used non-boss players will be able to see the respective team's icons in the top right of the player list of each team.