Builder buggeroff hotfix (#5214)
The loop kicks out units from the queue based on checking if
GetUnitCurrentCommand == nil, but I had placed the new code above that
point. A dead unit has a nil team so it can get caught in a crashing
loop.
Move the new code down into the main body so it doesn't trigger on dead
units.