Accurate reclaim value on large wrecks (#5246)
I saw a reclaim area's value not update for a long time on someone's
stream earlier. This is an issue when reclaiming commander wrecks,
mostly, since it's often ideal to leave a tiny amount of metal instead
of destroying the wreck.

The issue was a sequence-length check done on a hash map; the proper way
to check for entries in a hash in lua is `next(tbl)`.

### Work done

- update reclaim text at same rate as reclustering
- update reclaim text even when clustering doesn't occur

Co-authored-by: efrec <quantumscrewdriver@gmail.com>