Allow higher feature drawdistance (#5400)
Allow maximum featuredrawdistance in settings to go all the way to 40,000, which allows player to always see features even when fully zoomed out on the largest maps. Default value of 10,000 remains unchanged

The max value has not been increased for a few years at least while maps have gotten bigger and computers faster. I've had it manually set to 50,000 for some time now and it's never been an issue.