Prevent factory-assisting builders from leaving to go repair damaged finished units (#5441)
In Recoil engine, builders guarding factories generate a repair command for each unit they assist in building, appending it to the front of their command queue.

This causes issues if the unit comes out in a damaged state, in which case the builders will follow the new unit around and try to fix it before they return to guard.

Add new gadget that checks if a damaged unit is produced by a factory. If so, it removes the REPAIR command from any assisting builders.