Armwar script rework, to enable firing while turning without burst = 3 (#5532)
Large script rework of armwar. Original goal was to just implement
Friendly Fire mitigation code. But the change to a 3 burst on Jul 18,
2024 made that complicated, as Friendly Fire checks are not performed on
the 2nd and 3rd shots.

Original motivation for the Jul 18, 2024 change was to improve the
armwar script so that it would continue firing while turning/microed.
Therefore, the below script attempts to improve that script fix,
allowing the change to a 3 burst to be reverted, so that Friendly Fire
issues can be fixed.

Major change is shunting all turret turning logic to
ContinuousAimAccounting(), a function that is started on Create(),
instead of trying to do it all in AimWeapon().