Redefine "Utility Value" in the Spectator HUD widget (#5220)
Utility value in the Spectator HUD widget is not very useful. It only includes things like radars and jammers. These rarely exceed 2k metal in value, even for a full 8 player team, and being such a small value category means it is quite insignificant.

Redefine utility value to include the metal value of all units that are not included in the other value categories. Thus, every unit contributes value to one of the 4 value categories: army, defense, economy, or utility.