Fix: Initial spawn sets proper unit blocking (#5614)
Force blocked map squares to update by setting and un-setting unit blocking on units during the initial spawn script.

This change is a minimal change to fix blocking only, to be clear. But it could be used as a starting point to patch out the initial 3-second-ish pause at the start of the game, or to replace the current approach used to immobilize and paralyze the commander at spawn during its "warping in" time. MoveCtrl and paralysis may both be unnecessary semi-hacks that I can replace.